basic command diagnostic
